+++ src/optimum/rbln/transformers/models/whisper/whisper_architecture.py
@@ -300,7 +300,7 @@ class WhisperSelfAttention(WhisperAttention):
             "kcache": past_key_value[0].view(num_blocks, self.num_heads, 1, -1, self.head_dim),
             "vcache": past_key_value[1].view(num_blocks, self.num_heads, 1, -1, self.head_dim),
             "seq": cache_position.expand(bsz, 1),
-            "scale": torch.tensor(1.0, dtype=torch.float32),
+            "scale": torch.tensor(1.0, dtype=query_states.dtype),
             "block_table": block_tables,
             "block_size": block_size,
         }
